% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_table.R
\name{bottomright}
\alias{bottomright}
\alias{br}
\title{prints the bottomright corner of a 2D array (matrix, data.frame, data.table) to the console}
\usage{
bottomright(x, rcnt = 5, ccnt = rcnt)

br(x, rcnt = 5, ccnt = rcnt)
}
\arguments{
\item{x}{a 2D array}

\item{rcnt}{row count to print, default = 5}

\item{ccnt}{column count to print, default = rcnt}
}
\value{
nothing
}
\description{
prints the bottomright corner of a 2D array (matrix, data.frame, data.table) to the console
}
