% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_other.R
\name{catn}
\alias{catn}
\title{utility function to cat undefined number of arguments and end with trailing newline}
\usage{
catn(...)
}
\arguments{
\item{...}{strings to cat}
}
\description{
utility function to cat undefined number of arguments and end with trailing newline
}
