% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_io.R
\name{check_csv_region}
\alias{check_csv_region}
\title{identifies the local csv type used in the provided file}
\usage{
check_csv_region(csv, n_lines = 5)
}
\arguments{
\item{csv}{path to csv file}

\item{n_lines}{amount of lines to read to analyze the type of the file}
}
\value{
either "us" or "eu"
}
\description{
identifies the local csv type used in the provided file
}
