% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_io.R
\name{clean_path}
\alias{clean_path}
\title{noramlize path (get rid of ..) but default to forward slash}
\usage{
clean_path(path)
}
\arguments{
\item{path}{string path (e.g. result of file.path()}
}
\value{

}
\description{
noramlize path (get rid of ..) but default to forward slash
}
