% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_rstudio.R
\name{clear}
\alias{clear}
\title{\code{clear} will clear the global environment or, when specified, a list of
variables. \code{clear} will then call the garbage collector.}
\usage{
clear(x = c())
}
\arguments{
\item{x}{a vector of variable names to be deleted from the workspace}
}
\description{
\code{clear} will clear the global environment or, when specified, a list of
variables. \code{clear} will then call the garbage collector.
}
