% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_io.R
\name{clip_path}
\alias{clip_path}
\title{replaces backslashes with forward slashes in a path on the clipboard}
\usage{
clip_path()
}
\description{
replaces backslashes with forward slashes in a path on the clipboard
}
