% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_io.R
\name{combine_path}
\alias{combine_path}
\title{utility function to combine pieces of a path into a nice character path --> better: use file.path()}
\usage{
combine_path(...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
utility function to combine pieces of a path into a nice character path --> better: use file.path()
}
