% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_io.R
\name{convert_csv_regional_style}
\alias{convert_csv_regional_style}
\title{convert a CSV from one regional style to another}
\usage{
convert_csv_regional_style(fp_in, overwrite = FALSE, direction = "eu2us")
}
\arguments{
\item{fp_in}{string path for input csv file in EU notation}

\item{overwrite}{logical}

\item{direction}{currently supported choices are "eu2us" or "us2eu"}
}
\value{
output file path or NULL if error
}
\description{
convert a CSV from one regional style to another
}
