% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_datetime.R
\name{convert_datetime_to_timeid}
\alias{convert_datetime_to_timeid}
\title{provided a vector of string dates in the standard year 2018 it will return a
vector of timeids}
\usage{
convert_datetime_to_timeid(datetime)
}
\arguments{
\item{datetime}{string in the format ymd hm}
}
\value{

}
\description{
provided a vector of string dates in the standard year 2018 it will return a
vector of timeids
}
