% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_project.R
\name{count_lines_of_code}
\alias{count_lines_of_code}
\title{returns total lines of R code for all .R files in wd}
\usage{
count_lines_of_code(wd)
}
\arguments{
\item{wd}{string reference to a folder}
}
\description{
returns total lines of R code for all .R files in wd
}
