% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_vector.R
\name{csv_to_vector}
\alias{csv_to_vector}
\title{splits a comma separated character string into a character vector}
\usage{
csv_to_vector(char_string, split = ",")
}
\arguments{
\item{char_string}{character string}

\item{split}{separation character}
}
\value{

}
\description{
splits a comma separated character string into a character vector
}
