% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_table.R
\name{distribute_vector_over_matrix}
\alias{distribute_vector_over_matrix}
\title{distributes per row each element of m_distribute over the elements of 
m_data pro rata the weight of each element per line in m_data}
\usage{
distribute_vector_over_matrix(m_data, m_distribute)
}
\arguments{
\item{m_data}{x n numeric matrix}

\item{m_distribute}{m x 1 numeric vector}
}
\value{
new matrix
}
\description{
distributes per row each element of m_distribute over the elements of 
m_data pro rata the weight of each element per line in m_data
}
\examples{
m_distribute = c(1,10)
m_data = matrix(c(1,2,20,10), nrow = 2, byrow = TRUE)
distribute_vector_over_matrix(m_data, m_distribute)
}
