% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_other.R
\name{docvar}
\alias{docvar}
\title{cats a string representation of a variable's structure, 
and puts it on the clipboard if desired, to be pasted in roxygen doc}
\usage{
docvar(var, type = "return", cb = TRUE)
}
\arguments{
\item{var}{the variable you wish to document}

\item{type}{is it roxygen 'param' or 'return'?}

\item{cb}{TRUE FALSE : put doc on clipboard}
}
\value{
a string documenting the variable
}
\description{
cats a string representation of a variable's structure, 
and puts it on the clipboard if desired, to be pasted in roxygen doc
}
