% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_excel.R
\name{excel_cell_to_RC}
\alias{excel_cell_to_RC}
\title{convert an excel cell reference to a list of row and number integers}
\usage{
excel_cell_to_RC(s)
}
\arguments{
\item{s}{excel cell reference as string}
}
\value{
list(R = ..., C = ...)
}
\description{
convert an excel cell reference to a list of row and number integers
}
