% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_io.R
\name{file_backup}
\alias{file_backup}
\title{will create a backup of a file where the file is suffixed with a timestamp and '.backup'}
\usage{
file_backup(fpfn_from)
}
\arguments{
\item{fpfn_from}{file path to file to be backed up}
}
\value{

}
\description{
will create a backup of a file where the file is suffixed with a timestamp and '.backup'
}
