% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_vector.R
\name{fill_A_NA_with_B}
\alias{fill_A_NA_with_B}
\title{fill NA in vector A with values from vector B}
\usage{
fill_A_NA_with_B(a, b)
}
\arguments{
\item{a}{vector containing NA to replace}

\item{b}{vector containing substitute values, must be same length as a}
}
\value{
improved vector a
}
\description{
fill NA in vector A with values from vector B
}
\examples{
a = c(1,NA,3,NA)
b = 1:4
fill_A_NA_with_B(a,b) 
}
