% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_vector.R
\name{fractionalise}
\alias{fractionalise}
\title{returns a vector where every element is replaced by its fraction of the total sum of the vector}
\usage{
fractionalise(x)
}
\arguments{
\item{x}{a numerical vector}
}
\value{
a numerical vector containing fractional values (0 <= ... <= 1)
}
\description{
returns a vector where every element is replaced by its fraction of the total sum of the vector
}
