% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_datetime.R
\name{get_season_from_date}
\alias{get_season_from_date}
\title{returns the season string for a given date}
\usage{
get_season_from_date(input_date)
}
\arguments{
\item{input_date}{}
}
\value{
a string from the list: c("winter","inter","summer","inter","winter")
}
\description{
returns the season string for a given date
}
