% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_rstudio.R
\name{get_session_type}
\alias{get_session_type}
\title{Provides information about the type of sessions we are running}
\usage{
get_session_type()
}
\value{
a boolean 'interactive' = TRUE when code is run from RStudio, FALSE if it is run from RScript.exe
}
\description{
Provides information about the type of sessions we are running
}
