% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_io.R
\name{init_project_folder}
\alias{init_project_folder}
\title{given that the workdir is project/src this will create project/foldername if it does not exist yet}
\usage{
init_project_folder(foldername)
}
\arguments{
\item{foldername}{}
}
\value{
the full file path of the folder you wanted to initialise
}
\description{
given that the workdir is project/src this will create project/foldername if it does not exist yet
}
