% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_vector.R
\name{insert_at_element}
\alias{insert_at_element}
\title{insert a vector of elements in a vector before or after a given element}
\usage{
insert_at_element(v, existing_element, new_elements, where = "before")
}
\arguments{
\item{v}{existing vector}

\item{existing_element}{positional element, should appear exactly once in the vector}

\item{new_elements}{vector of new elements}

\item{where}{"before" or "after"}
}
\value{
a vector with the new values merged in at the correct location
}
\description{
insert a vector of elements in a vector before or after a given element
}
\examples{
insert_at_element(c("a","b","e"), "a", c("c","d"), "before")
insert_at_element(c("a","b","e"), "a", c("c","d"), "after")
insert_at_element(c("a","b","e"), "b", c("c","d"), "before")
insert_at_element(c("a","b","e"), "b", c("c","d"), "after")
insert_at_element(c("a","b","e"), "e", c("c","d"), "before")
}
