% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_excel.R
\name{letters_to_numbers}
\alias{letters_to_numbers}
\title{convert excel letter column reference to numeric}
\usage{
letters_to_numbers(s)
}
\arguments{
\item{s}{excel column reference as string (e.g. A, AB, DYX, ...)}
}
\value{

}
\description{
convert excel letter column reference to numeric
}
