% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_other.R
\name{lines_of_code_per_file}
\alias{lines_of_code_per_file}
\title{count lines of code in R & python files in your project}
\usage{
lines_of_code_per_file(files, skip_comments = FALSE, skip_blanks = FALSE)
}
\arguments{
\item{files}{vector of file paths to count - see list_code_files()}

\item{skip_comments}{does not count lines starting in '#'}

\item{skip_blanks}{does not count blank lines (whitespace)}
}
\value{
integer lines of code
}
\description{
count lines of code in R & python files in your project
}
