% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_other.R
\name{list_code_files}
\alias{list_code_files}
\title{returns a character vector of absolute code file paths}
\usage{
list_code_files(path, ext = c("py", "r"), skiplist = c("/renv/", "/venv/"))
}
\arguments{
\item{path}{parent folder to look through recursively}

\item{ext}{character vector of extensions to accept (case insensitive, no dot)}

\item{skiplist}{character vector of pieces of path to skip}
}
\value{
character vector of code file paths
}
\description{
returns a character vector of absolute code file paths
}
