% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_table.R
\name{list_of_vectors_to_dt}
\alias{list_of_vectors_to_dt}
\title{creates a new data.table where every list element which is a vector becomes a column}
\usage{
list_of_vectors_to_dt(l)
}
\arguments{
\item{l}{list of vectors of length N, or dataframes with nrow = N}
}
\value{

}
\description{
creates a new data.table where every list element which is a vector becomes a column
}
