% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_other.R
\name{load_packages_robustly}
\alias{load_packages_robustly}
\title{Test whether the provided vector of package is installed on the users's system
Installed packages are loaded. The user is prompted to install non-installed packages.}
\usage{
load_packages_robustly(packages, prompt = TRUE)
}
\arguments{
\item{packages}{Character vector of pkg names}

\item{prompt}{Boolean whether the user should be prompted to install missing packages}
}
\value{
Boolean indicating if all packages are installed
}
\description{
Test whether the provided vector of package is installed on the users's system
Installed packages are loaded. The user is prompted to install non-installed packages.
}
