% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_other.R
\name{loc}
\alias{loc}
\title{shortcut for list_code_file and count_lines_of_code that defaults to including comments and excluding blanks}
\usage{
loc(path)
}
\arguments{
\item{path}{root path of a project}
}
\value{
integer lines of code
}
\description{
shortcut for list_code_file and count_lines_of_code that defaults to including comments and excluding blanks
}
