% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_list.R
\name{merge_lists_by_key}
\alias{merge_lists_by_key}
\title{merge 2 lists of lists together by common key}
\usage{
merge_lists_by_key(list1, list2)
}
\arguments{
\item{list1}{first list}

\item{list2}{second list}
}
\value{
a merged list where elements of common key are concatenated into a new sublist
}
\description{
merge 2 lists of lists together by common key
}
