% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_table.R
\name{merge_overwrite}
\alias{merge_overwrite}
\title{merge a data.table x and y (left join) and overwrite the 
values in x with those in y for the matching columns}
\usage{
merge_overwrite(x, y, by = NULL, by.x = NULL, by.y = NULL)
}
\arguments{
\item{x}{data.table}

\item{y}{data.table}

\item{by}{A vector of shared column names in x and y to merge on. This defaults to the shared key columns between the two tables. If y has no key columns, this defaults to the key of x}

\item{by.x}{Vectors of column names in x and y to merge on.}

\item{by.y}{Vectors of column names in x and y to merge on.}
}
\value{

}
\description{
merge a data.table x and y (left join) and overwrite the 
values in x with those in y for the matching columns
}
