% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_conversions.R
\name{named_list}
\alias{named_list}
\title{create a list where for each element the name of the variable is the key and the content is the variable content}
\usage{
named_list(...)
}
\arguments{
\item{...}{comma seperated list of variables}
}
\value{
named list
}
\description{
create a list where for each element the name of the variable is the key and the content is the variable content
}
\examples{
a <- b <- c <- 1
namedList(a,b,c)
}
