% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_vector.R
\name{nearest_pos}
\alias{nearest_pos}
\title{looks for the value closest to the provided values, in a vector of provided values, return the position}
\usage{
nearest_pos(look_near_these_values, in_these_values)
}
\arguments{
\item{look_near_these_values}{look for values close to these}

\item{in_these_values}{look in this vector}
}
\value{
the position of the closest values
}
\description{
looks for the value closest to the provided values, in a vector of provided values, return the position
}
