% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_vector.R
\name{nearest_val}
\alias{nearest_val}
\title{looks for the value closest to the provided values, in a vector of provided values, return the value}
\usage{
nearest_val(look_near_these_values, in_these_values)
}
\arguments{
\item{look_near_these_values}{look for values close to these}

\item{in_these_values}{look in this vector}
}
\value{

}
\description{
looks for the value closest to the provided values, in a vector of provided values, return the value
}
