% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_datetime.R
\name{now}
\alias{now}
\title{return a string timestamp in ISO8601 using the provided separators}
\usage{
now(date_hour_sep = "-", date_sep = "", hour_sep = "")
}
\arguments{
\item{date_hour_sep}{separator to be used between date and hour segment of the output}

\item{date_sep}{separator to be used between date components of the output}

\item{hour_sep}{separator to be used between hour components of the output}
}
\value{
string timestamp
}
\description{
return a string timestamp in ISO8601 using the provided separators
}
