% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_excel.R
\name{numeric_date_to_iso8601}
\alias{numeric_date_to_iso8601}
\title{converts a numeric date (e.g. when reading an excel file) to a iso8601 date string
works on windows only}
\usage{
numeric_date_to_iso8601(num)
}
\arguments{
\item{num}{numeric representation of date - for Excel on Windows, the origin date is December 30, 1899 for dates after 1900.}
}
\value{

}
\description{
converts a numeric date (e.g. when reading an excel file) to a iso8601 date string
works on windows only
}
