% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_rstudio.R
\name{print_globalenv_size}
\alias{print_globalenv_size}
\title{returns the size of the global workspace in a human readable format}
\usage{
print_globalenv_size()
}
\value{
the memory usage of the global environment
}
\description{
returns the size of the global workspace in a human readable format
}
