% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_rstudio.R
\name{print_var_size}
\alias{print_var_size}
\title{returns the size of a single variable in a human readable format}
\usage{
print_var_size(variableName)
}
\arguments{
\item{variableName}{name of the variable for which to display the memory usage.}
}
\value{
the size of the variable
}
\description{
returns the size of a single variable in a human readable format
}
