% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_project.R
\name{projects_lines_count}
\alias{projects_lines_count}
\title{returns the number of R code lines in current working directory}
\usage{
projects_lines_count()
}
\value{
numeric
}
\description{
returns the number of R code lines in current working directory
}
