% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_other.R
\name{pt}
\alias{pt}
\title{create a progress string like "(5/10)"}
\usage{
pt(i, tot, tspace = TRUE)
}
\arguments{
\item{i}{numeric}

\item{tot}{numeric}

\item{tspace}{boolean for trailing space}
}
\value{

}
\description{
create a progress string like "(5/10)"
}
