% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_string.R
\name{remove_all_spec_char}
\alias{remove_all_spec_char}
\title{removes all non alphanumerical characters except - and _ from a string}
\usage{
remove_all_spec_char(s, repl = "_")
}
\arguments{
\item{s}{input string}

\item{repl}{replacement character}
}
\value{
string with special characters removed
}
\description{
removes all non alphanumerical characters except - and _ from a string
}
