% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_table.R
\name{remove_tail_na}
\alias{remove_tail_na}
\title{removes tail rows that contain NA values, up until the last row that is fully complete}
\usage{
remove_tail_na(dt)
}
\arguments{
\item{dt}{data.frame}
}
\value{

}
\description{
removes tail rows that contain NA values, up until the last row that is fully complete
}
