% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_vector.R
\name{rescale}
\alias{rescale}
\title{rescale a numerical vector using a new total sum value}
\usage{
rescale(x, total)
}
\arguments{
\item{x}{a numerica lvector}

\item{total}{the new total sum to scale to}
}
\value{
a rescaled numerical vector with total sum = total argument
}
\description{
rescale a numerical vector using a new total sum value
}
