% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_string.R
\name{str_left}
\alias{str_left}
\title{show leftmost x characters from string s}
\usage{
str_left(s, x)
}
\arguments{
\item{s}{string}

\item{x}{integer}
}
\value{
string
}
\description{
show leftmost x characters from string s
}
