% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_string.R
\name{str_right}
\alias{str_right}
\title{show rightmost x characters from string sTitle}
\usage{
str_right(s, x)
}
\arguments{
\item{s}{string}

\item{x}{integer}
}
\value{
string
}
\description{
show rightmost x characters from string sTitle
}
