% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_conversions.R
\name{string_gently_to_boolean}
\alias{string_gently_to_boolean}
\title{converts a string to a boolean if it holds true or false, but retains the string if not}
\usage{
string_gently_to_boolean(x)
}
\arguments{
\item{x}{string "TRUE", "true", "FALSE", "false" or something else}
}
\value{

}
\description{
converts a string to a boolean if it holds true or false, but retains the string if not
}
