% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_conversions.R
\name{string_gently_to_numeric}
\alias{string_gently_to_numeric}
\title{converts a string to a numeric if it holds a number, but retains the string if not}
\usage{
string_gently_to_numeric(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
converts a string to a numeric if it holds a number, but retains the string if not
}
