% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_other.R
\name{tik}
\alias{tik}
\title{utility function to cat undefined number of arguments and end with trailing newline 
and leading timestamp, and starting tic()}
\usage{
tik(...)
}
\arguments{
\item{...}{strings to cat}
}
\description{
utility function to cat undefined number of arguments and end with trailing newline 
and leading timestamp, and starting tic()
}
