% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_table.R
\name{topleft}
\alias{topleft}
\alias{tl}
\title{prints the topleft corner of a 2D array (matrix, data.frame, data.table) to the console}
\usage{
topleft(x, rcnt = 5, ccnt = rcnt)

tl(x, rcnt = 5, ccnt = rcnt)
}
\arguments{
\item{x}{a 2D array}

\item{rcnt}{row count to print, default = 5}

\item{ccnt}{column count to print, default = rcnt}
}
\value{
nothing
}
\description{
prints the topleft corner of a 2D array (matrix, data.frame, data.table) to the console
}
