% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_table.R
\name{topright}
\alias{topright}
\alias{tr}
\title{prints the topright corner of a 2D array (matrix, data.frame, data.table) to the console}
\usage{
topright(x, rcnt = 5, ccnt = rcnt)

tr(x, rcnt = 5, ccnt = rcnt)
}
\arguments{
\item{x}{a 2D array}

\item{rcnt}{row count to print, default = 5}

\item{ccnt}{column count to print, default = rcnt}
}
\value{
nothing
}
\description{
prints the topright corner of a 2D array (matrix, data.frame, data.table) to the console
}
