% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_datetime.R
\name{tstampnum}
\alias{tstampnum}
\title{version of tstamp without any special characters (only numerical characters) so that it can be used in filenames etc.}
\usage{
tstampnum()
}
\value{
numerical timestamp
}
\description{
version of tstamp without any special characters (only numerical characters) so that it can be used in filenames etc.
}
