% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_io.R
\name{two_column_csv_to_list}
\alias{two_column_csv_to_list}
\title{parses a 2 column csv file into a list, usefull for options csv's}
\usage{
two_column_csv_to_list(csv_fpfn)
}
\arguments{
\item{csv_fpfn}{flie path to csv file}
}
\value{

}
\description{
parses a 2 column csv file into a list, usefull for options csv's
}
