% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_io.R
\name{two_column_xlsx_to_list}
\alias{two_column_xlsx_to_list}
\title{parses a 2 column xlsx file into a list, usefull for options files}
\usage{
two_column_xlsx_to_list(xlsx_fpfn, xlsx_sheet)
}
\arguments{
\item{xlsx_fpfn}{xlsx file path}

\item{xlsx_sheet}{string name of sheet or integer number of sheet to read}
}
\value{
a list object with settings
}
\description{
parses a 2 column xlsx file into a list, usefull for options files
}
