% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_other.R
\name{update_jrutils_package}
\alias{update_jrutils_package}
\title{run this to update the sma_r_utils package from a source bundle}
\usage{
update_jrutils_package()
}
\description{
run this to update the sma_r_utils package from a source bundle
}
