% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_vector.R
\name{vector_to_csv}
\alias{vector_to_csv}
\title{returns a character vector as a comma seperated string}
\usage{
vector_to_csv(char_vector, split = ",")
}
\arguments{
\item{char_vector}{character vector}

\item{split}{separation character}
}
\value{

}
\description{
returns a character vector as a comma seperated string
}
