% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_io.R
\name{wopen}
\alias{wopen}
\title{utility function to open path in windows explorer}
\usage{
wopen(path)
}
\arguments{
\item{path}{string to folderpath}
}
\description{
utility function to open path in windows explorer
}
