% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_io.R
\name{write_table_to_clipboard}
\alias{write_table_to_clipboard}
\title{shortcut to put your matrix, data.frame or data.table on your clipboard.}
\usage{
write_table_to_clipboard(x, size = 1024)
}
\arguments{
\item{x}{R table structure (data.frame, matrix, ...)}

\item{size}{extend clipboard size if data.table is too large}
}
\value{

}
\description{
shortcut to put your matrix, data.frame or data.table on your clipboard.
}
