% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_other.R
\name{loc}
\alias{loc}
\title{count lines of code in R & python files in your project}
\usage{
loc(path, skip_comments = FALSE, skip_blanks = TRUE)
}
\arguments{
\item{path}{root path of a project}

\item{skip_comments}{does not count lines starting in '#'}

\item{skip_blanks}{does not count blank lines (whitespace)}
}
\value{
integer lines of code
}
\description{
count lines of code in R & python files in your project
}
