% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_io.R
\name{robust_fread}
\alias{robust_fread}
\title{fread should be able to auto detect CSV seperator, this function makes sure
it does so always}
\usage{
robust_fread(fpfn, header = "auto", skip = 0)
}
\arguments{
\item{fpfn}{filepath to csv file}

\item{header}{TRUE/FALSE to identify if your file has a header}

\item{skip}{integer amount of rows to skip}
}
\value{
csv file contents as data.table
}
\description{
fread should be able to auto detect CSV seperator, this function makes sure
it does so always
}
